+++ apps/bare-expo/ios/Pods/Headers/Private/CocoaLibEvent/event2/rpc_compat.h
@@ -1,1 +1,0 @@
-../../../../CocoaLibEvent/src/event2/rpc_compat.h