+++ apps/bare-expo/ios/Pods/Headers/Private/Flipper-Folly/folly/stop_watch.h
@@ -1,1 +1,0 @@
-../../../../Flipper-Folly/folly/stop_watch.h