+++ apps/bare-expo/ios/Pods/Headers/Private/ZXingObjC/ZXUPCAReader.h
@@ -1,1 +1,0 @@
-../../../ZXingObjC/ZXingObjC/oned/ZXUPCAReader.h