+++ apps/bare-expo/ios/Pods/Headers/Public/CocoaLibEvent/event2/http.h
@@ -1,1 +1,0 @@
-../../../../CocoaLibEvent/src/event2/http.h