+++ apps/bare-expo/ios/Pods/Headers/Public/EXTaskManager/EXTaskService.h
@@ -1,1 +1,0 @@
-../../../../../../../packages/expo-task-manager/ios/EXTaskManager/EXTaskService.h