+++ apps/bare-expo/ios/Pods/Headers/Public/Flipper/CertificateUtils.h
@@ -1,1 +1,0 @@
-../../../Flipper/xplat/Flipper/CertificateUtils.h