+++ apps/bare-expo/ios/Pods/Headers/Public/RNSVG/RNSVGBezierElement.h
@@ -1,1 +1,0 @@
-../../../../../../../node_modules/react-native-svg/ios/Utils/RNSVGBezierElement.h