+++ apps/bare-expo/ios/Pods/Headers/Public/React-jsiexecutor/jsireact/JSINativeModules.h
@@ -1,1 +1,0 @@
-../../../../../../node_modules/react-native/ReactCommon/jsiexecutor/jsireact/JSINativeModules.h