+++ apps/bare-expo/ios/Pods/Target Support Files/Pods-BareExpoMain-BareExpo/Pods-BareExpoMain-BareExpo-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Pods_BareExpoMain_BareExpo : NSObject
-@end
-@implementation PodsDummy_Pods_BareExpoMain_BareExpo
-@end