+++ apps/bare-expo/ios/Pods/Target Support Files/RNSharedElement/RNSharedElement-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_RNSharedElement : NSObject
-@end
-@implementation PodsDummy_RNSharedElement
-@end